# backend/generated/openapi-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/treasury/tariffs": {
      "get": {
        "operationId": "TreasuryController_getTariffs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TariffResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "List all active school tariffs",
        "tags": [
          "treasury"
        ]
      },
      "post": {
        "operationId": "TreasuryController_createTariff",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTariffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffResponse"
                }
              }
            }
          }
        },
        "summary": "Create a new school tariff",
        "tags": [
          "treasury"
        ]
      }
    },
    "/api/treasury/tariffs/{id}": {
      "put": {
        "operationId": "TreasuryController_updateTariff",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTariffDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffResponse"
                }
              }
            }
          }
        },
        "summary": "Update an existing tariff",
        "tags": [
          "treasury"
        ]
      },
      "delete": {
        "operationId": "TreasuryController_deleteTariff",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a tariff concept",
        "tags": [
          "treasury"
        ]
      }
    },
    "/api/treasury/charges": {
      "get": {
        "operationId": "TreasuryController_getCharges",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChargeResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "List all outstanding student charges",
        "tags": [
          "treasury"
        ]
      },
      "post": {
        "operationId": "TreasuryController_createCharge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChargeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChargeResponse"
                }
              }
            }
          }
        },
        "summary": "Create a direct charge for a student",
        "tags": [
          "treasury"
        ]
      }
    },
    "/api/treasury/charges/bulk": {
      "post": {
        "operationId": "TreasuryController_generateBulkCharges",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateBulkChargesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Number of charges successfully created."
          }
        },
        "summary": "Generate monthly bulk charges",
        "tags": [
          "treasury"
        ]
      }
    },
    "/api/treasury/charges/{id}": {
      "delete": {
        "operationId": "TreasuryController_deleteCharge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a pending charge",
        "tags": [
          "treasury"
        ]
      }
    },
    "/api/treasury/payments": {
      "get": {
        "operationId": "TreasuryController_getPayments",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chargeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "List recent cash/bank transaction payments",
        "tags": [
          "treasury"
        ]
      },
      "post": {
        "operationId": "TreasuryController_registerPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          }
        },
        "summary": "Register a payment receipt against a charge",
        "tags": [
          "treasury"
        ]
      }
    },
...
  },
  "info": {
    "title": "School Management API",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "apiKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apiKeyCookie",
        "description": "API Key authentication via cookie"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication"
      }
    },
    "schemas": {
...
      "SectionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de la sección",
            "example": "section-123"
          },
          "name": {
            "type": "string",
            "description": "Nombre (ej. A, B, C)",
            "example": "A"
          },
          "grade": {
            "type": "string",
            "description": "Grado",
            "example": "1ro de Primaria"
          },
          "level": {
            "type": "string",
            "enum": [
              "INITIAL",
              "PRIMARY",
              "SECONDARY"
            ],
            "example": "PRIMARY"
          },
          "capacity": {
            "type": "number",
            "description": "Aforo máximo permitido",
            "example": 30
          },
          "status": {
            "type": "string",
            "description": "Estado (ej. OPEN, CLOSED)",
            "example": "OPEN"
          },
          "matriculados": {
            "type": "number",
            "description": "Cantidad actual de alumnos matriculados",
            "example": 12
          }
        },
        "required": [
          "id",
          "name",
          "grade",
          "level",
          "capacity",
          "status",
          "matriculados"
        ]
      },
      "StudentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único del estudiante",
            "example": "student-123"
          },
          "code": {
            "type": "string",
            "description": "Código de estudiante",
            "example": "ALU-2026-1234",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "Nombres",
            "example": "Luis"
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos",
            "example": "Pérez"
          },
          "dni": {
            "type": "string",
            "description": "DNI",
            "example": "87654321"
          },
          "level": {
            "type": "string",
            "enum": [
              "INITIAL",
              "PRIMARY",
              "SECONDARY"
            ],
            "example": "PRIMARY"
          },
          "grade": {
            "type": "string",
            "description": "Grado",
            "example": "1ro de Primaria"
          },
          "sectionId": {
            "type": "string",
            "description": "ID de la sección",
            "example": "section-123",
            "nullable": true
          },
          "section": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/SectionResponse"
              }
            ]
          },
          "guardianId": {
            "type": "string",
            "description": "ID del apoderado",
            "example": "guardian-123"
          },
          "guardian": {
            "$ref": "#/components/schemas/GuardianResponse"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de creación"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "dni",
          "level",
          "grade",
          "guardianId",
          "createdAt"
        ]
      },
      "GuardianResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único del apoderado",
            "example": "guardian-123"
          },
          "dni": {
            "type": "string",
            "description": "DNI del apoderado",
            "example": "12345678"
          },
          "name": {
            "type": "string",
            "description": "Nombre completo",
            "example": "Juan Pérez"
          },
          "phone": {
            "type": "string",
            "description": "Teléfono de contacto",
            "example": "987654321"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico",
            "example": "juan.perez@gmail.com",
            "nullable": true
          },
          "occupation": {
            "type": "string",
            "description": "Ocupación",
            "example": "Ingeniero",
            "nullable": true
          },
          "students": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentResponse"
            }
          }
        },
        "required": [
          "id",
          "dni",
          "name",
          "phone"
        ]
      },
...
      "TariffResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de la tarifa",
            "example": "tariff-123"
          },
          "concept": {
            "type": "string",
            "description": "Concepto de la tarifa",
            "example": "Pensión Mensual"
          },
          "amount": {
            "type": "number",
            "description": "Monto de la tarifa",
            "example": 450
          },
          "type": {
            "type": "string",
            "enum": [
              "ONE_TIME",
              "MONTHLY",
              "EXTRA"
            ],
            "description": "Tipo de cobro de la tarifa"
          },
          "level": {
            "type": "string",
            "enum": [
              "INITIAL",
              "PRIMARY",
              "SECONDARY"
            ],
            "description": "Nivel educativo aplicable"
          }
        },
        "required": [
          "id",
          "concept",
          "amount",
          "type",
          "level"
        ]
      },
      "CreateTariffDto": {
        "type": "object",
        "properties": {
          "concept": {
            "type": "string",
            "description": "Concepto de la tarifa",
            "example": "Pensión Mensual de Mayo"
          },
          "amount": {
            "type": "number",
            "description": "Monto de la tarifa",
            "example": 450
          },
          "type": {
            "type": "string",
            "enum": [
              "ONE_TIME",
              "MONTHLY",
              "EXTRA"
            ],
            "description": "Tipo de cobro de la tarifa",
            "example": "MONTHLY"
          },
          "level": {
            "type": "string",
            "enum": [
              "INITIAL",
              "PRIMARY",
              "SECONDARY"
            ],
            "description": "Nivel educativo aplicable",
            "example": "PRIMARY"
          }
        },
        "required": [
          "concept",
          "amount",
          "type",
          "level"
        ]
      },
      "UpdateTariffDto": {
        "type": "object",
        "properties": {
          "concept": {
            "type": "string",
            "description": "Concepto de la tarifa",
            "example": "Pensión Mayo Modificada"
          },
          "amount": {
            "type": "number",
            "description": "Monto de la tarifa",
            "example": 480
          },
          "type": {
            "type": "string",
            "enum": [
              "ONE_TIME",
              "MONTHLY",
              "EXTRA"
            ],
            "description": "Tipo de cobro de la tarifa",
            "example": "MONTHLY"
          },
          "level": {
            "type": "string",
            "enum": [
              "INITIAL",
              "PRIMARY",
              "SECONDARY"
            ],
            "description": "Nivel educativo aplicable",
            "example": "PRIMARY"
          }
        }
      },
      "ChargeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del cargo",
            "example": "charge-123"
          },
          "studentId": {
            "type": "string",
            "description": "ID del estudiante",
            "example": "student-123"
          },
          "student": {
            "$ref": "#/components/schemas/StudentResponse"
          },
          "tariffId": {
            "type": "string",
            "description": "ID de la tarifa",
            "example": "tariff-123"
          },
          "tariff": {
            "$ref": "#/components/schemas/TariffResponse"
          },
          "originalAmount": {
            "type": "number",
            "description": "Monto original",
            "example": 450
          },
          "pendingAmount": {
            "type": "number",
            "description": "Monto pendiente",
            "example": 450
          },
          "dueDate": {
            "type": "object",
            "description": "Fecha de vencimiento",
            "example": "2026-06-05T00:00:00.000Z",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "Estado del cargo",
            "example": "PENDING"
          }
        },
        "required": [
          "id",
          "studentId",
          "tariffId",
          "originalAmount",
          "pendingAmount",
          "dueDate",
          "status"
        ]
      },
      "CreateChargeDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "description": "ID del estudiante",
            "example": "student-123"
          },
          "tariffId": {
            "type": "string",
            "description": "ID de la tarifa",
            "example": "tariff-123"
          },
          "dueDate": {
            "type": "string",
            "description": "Fecha de vencimiento",
            "example": "2026-06-05T00:00:00.000Z"
          }
        },
        "required": [
          "studentId",
          "tariffId"
        ]
      },
      "GenerateBulkChargesDto": {
        "type": "object",
        "properties": {
          "tariffId": {
            "type": "string",
            "description": "ID de la tarifa a aplicar",
            "example": "tariff-123"
          },
          "dueDate": {
            "type": "string",
            "description": "Fecha de vencimiento para los cargos",
            "example": "2026-06-05T00:00:00.000Z"
          }
        },
        "required": [
          "tariffId"
        ]
      },
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de la transacción de pago",
            "example": "payment-123"
          },
          "chargeId": {
            "type": "string",
            "description": "ID del cargo asociado",
            "example": "charge-123"
          },
          "charge": {
            "$ref": "#/components/schemas/ChargeResponse"
          },
          "totalAmount": {
            "type": "number",
            "description": "Monto pagado",
            "example": 150
          },
          "method": {
            "type": "string",
            "enum": [
              "CASH",
              "CARD",
              "TRANSFER"
            ],
            "description": "Método de pago"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha y hora de pago",
            "example": "2026-05-25T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "chargeId",
          "totalAmount",
          "method",
          "timestamp"
        ]
      },
      "RegisterPaymentDto": {
        "type": "object",
        "properties": {
          "chargeId": {
            "type": "string",
            "description": "ID del cargo a pagar",
            "example": "charge-123"
          },
          "amount": {
            "type": "number",
            "description": "Monto total a pagar",
            "example": 150
          },
          "method": {
            "type": "string",
            "enum": [
              "CASH",
              "CARD",
              "TRANSFER"
            ],
            "description": "Método de pago",
            "example": "CASH"
          }
        },
        "required": [
          "chargeId",
          "amount",
          "method"
        ]
      },
...
    }
  }
}